requirements.txt
python-telegram-bot==21.1.1
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.9.5
